+++ pyccel/codegen/printing/cwrappercode.py
@@ -8,12 +8,14 @@ from pyccel.codegen.printing.ccode import CCodePrinter
 
 from pyccel.ast.bind_c     import BindCPointer
 from pyccel.ast.bind_c     import BindCModule, BindCFunctionDef
+from pyccel.ast.c_concepts import CStackArray
 from pyccel.ast.core       import FunctionAddress, SeparatorComment
 from pyccel.ast.core       import Import, Module, Declare
-from pyccel.ast.cwrapper   import PyBuildValueNode
+from pyccel.ast.cwrapper   import PyBuildValueNode, PyCapsule_New, PyCapsule_Import, PyModule_Create
 from pyccel.ast.cwrapper   import Py_None, WrapperCustomDataType
 from pyccel.ast.cwrapper   import PyccelPyObject, PyccelPyTypeObject
-from pyccel.ast.literals   import LiteralString, Nil
+from pyccel.ast.literals   import LiteralString, Nil, LiteralInteger
+from pyccel.ast.numpy_wrapper import PyccelPyArrayObject
 from pyccel.ast.c_concepts import ObjectAddress
 
 from pyccel.errors.errors  import Errors
@@ -25,8 +27,6 @@ errors = Errors()
 module_imports = [Import('numpy_version', Module('numpy_version',(),())),
             Import('numpy/arrayobject', Module('numpy/arrayobject',(),())),
             Import('cwrapper', Module('cwrapper',(),()))]
-
-cwrapper_ndarray_import = Import('cwrapper_ndarrays', Module('cwrapper_ndarrays', (), ()))
 
 
 class CWrapperCodePrinter(CCodePrinter):
@@ -50,6 +50,7 @@ class CWrapperCodePrinter(CCodePrinter):
     """
     dtype_registry = {**CCodePrinter.dtype_registry,
                       (PyccelPyObject() , 0) : 'PyObject',
+                      (PyccelPyArrayObject() , 0) : 'PyArrayObject',
                       (PyccelPyTypeObject() , 0) : 'PyTypeObject',
                       (BindCPointer()   , 0) : 'void'}
 
@@ -85,9 +86,9 @@ class CWrapperCodePrinter(CCodePrinter):
         --------
         CCodePrinter.is_c_pointer : The extended function.
         """
-        if isinstance(a.dtype, (PyccelPyObject, PyccelPyTypeObject, WrapperCustomDataType)):
+        if isinstance(a.dtype, (WrapperCustomDataType, BindCPointer)):
             return True
-        elif isinstance(a, PyBuildValueNode):
+        elif isinstance(a, (PyBuildValueNode, PyCapsule_New, PyCapsule_Import, PyModule_Create)):
             return True
         else:
             return CCodePrinter.is_c_pointer(self,a)
@@ -245,11 +246,70 @@ class CWrapperCodePrinter(CCodePrinter):
             var = f'(PyObject*) {var}'
         return f'PyModule_AddObject({expr.mod_name}, {name}, {var})'
 
+    def _print_PyCapsule_New(self, expr):
+        name = expr.capsule_name
+        var  = self._print(ObjectAddress(expr.API_var))
+        return f'PyCapsule_New((void *){var}, "{name}", NULL)'
+
+    def _print_PyCapsule_Import(self, expr):
+        name = expr.capsule_name
+        return f'(void**)PyCapsule_Import("{name}", 0)'
+
+    def _print_PyModule_Create(self, expr):
+        return f'PyModule_Create(&{expr.module_def_name})'
+
+    def _print_ModuleHeader(self, expr):
+        mod = expr.module
+        name = mod.name
+
+        # Print imports last to be sure that all additional_imports have been collected
+        imports  = [*module_imports, *mod.imports]
+        for i in imports:
+            self.add_import(i)
+        imports  = ''.join(self._print(i) for i in imports)
+
+        function_signatures = ''.join(self.function_signature(f, print_arg_names = False) + ';\n' for f in mod.external_funcs)
+
+        API_var = mod.variables[0]
+
+        macro_defs = ''
+        classes = []
+        for i,c in enumerate(mod.classes):
+            struct_name = c.struct_name
+            type_name = c.type_name
+            attributes = ''.join(self._print(Declare(a.dtype, a)) for a in c.attributes)
+            classes.append(f"struct {struct_name} {{\n"
+                    "    PyObject_HEAD\n"
+                    + attributes +
+                    "};\n")
+            sig_methods = c.methods + (c.new_func,) + tuple(f for i in c.interfaces for f in i.functions) + \
+                          tuple(i.interface_func for i in c.interfaces)
+            function_signatures += '\n'+''.join(self.function_signature(f)+';\n' for f in sig_methods)
+            macro_defs += f'#define {type_name} (*(PyTypeObject*){API_var.name}[{i}])\n'
+
+        class_code = '\n'.join(classes)
+
+        static_import_decs = self._print(Declare(API_var.dtype, API_var, static=True))
+        import_func = self._print(mod.import_func)
+
+        header_id = f'{name.upper()}_WRAPPER'
+        header_guard = f'{header_id}_H'
+        return (f"#ifndef {header_guard}\n \
+                #define {header_guard}\n\n \
+                {imports}\n \
+                {class_code}\n \
+                #ifdef {header_id}\n\n \
+                {function_signatures}\n \
+                #else\n\n \
+                {static_import_decs}\n \
+                {macro_defs}\n \
+                {import_func}\n \
+                #endif\n \
+                #endif // {header_guard}\n")
+
     def _print_PyModule(self, expr):
         scope = expr.scope
         self.set_scope(scope)
-        # The initialisation and deallocation shouldn't be exposed to python
-        funcs_to_wrap = [f for f in expr.funcs if f not in (expr.init_func, expr.free_func)]
 
         # Insert declared objects into scope
         variables = expr.original_module.variables if isinstance(expr, BindCModule) else expr.variables
@@ -264,10 +324,8 @@ class CWrapperCodePrinter(CCodePrinter):
         self._module_name  = self.get_python_name(scope, expr)
         sep = self._print(SeparatorComment(40))
 
-        function_signatures = ''.join(self.function_signature(f, print_arg_names = False) + ';\n' for f in expr.external_funcs)
-
         interface_funcs = [f.name for i in expr.interfaces for f in i.functions]
-        funcs += [*expr.interfaces, *(f for f in funcs_to_wrap if f.name not in interface_funcs)]
+        funcs += [*expr.interfaces, *(f for f in expr.funcs if f.name not in interface_funcs)]
 
         self._in_header = True
         decs = ''.join(self._print(d) for d in expr.declarations)
@@ -287,14 +345,7 @@ class CWrapperCodePrinter(CCodePrinter):
                                             wrapper_name = f.name,
                                             docstring = self._print(LiteralString('\n'.join(f.docstring.comments))) \
                                                         if f.docstring else '""')
-                                     for f in funcs if f is not expr.init_func and not getattr(f, 'is_header', False))
-
-        slots_name = self.scope.get_new_name('{}_slots'.format(expr.name))
-        exec_func_name = expr.init_func.name
-        slots_def = (f'static PyModuleDef_Slot {slots_name}[] = {{\n'
-                     f'{{Py_mod_exec, {exec_func_name}}},\n'
-                     '{0, NULL},\n'
-                     '};\n')
+                                     for f in funcs if not getattr(f, 'is_header', False))
 
         method_def_name = self.scope.get_new_name('{}_methods'.format(expr.name))
         method_def = (f'static PyMethodDef {method_def_name}[] = {{\n'
@@ -312,28 +363,20 @@ class CWrapperCodePrinter(CCodePrinter):
                 '/* size of per-interpreter state of the module, or -1 if the module keeps state in global variables. */\n'
                 '0,\n'
                 f'{method_def_name},\n'
-                f'{slots_name}\n'
                 '};\n')
 
-        exec_func = self._print(expr.init_func)
+        init_func = self._print(expr.init_func)
 
-        init_func = (f'PyMODINIT_FUNC PyInit_{self._module_name}(void)\n{{\n'
-                'import_array();\n'
-                f'return PyModuleDef_Init(&{module_def_name});\n'
-                '}\n')
-
-        # Print imports last to be sure that all additional_imports have been collected
-        for i in expr.imports:
-            self.add_import(i)
-        imports  = module_imports.copy()
-        imports += self._additional_imports.values()
+        pymod_name = f'{expr.name}_wrapper'
+        imports = [Import(pymod_name, Module(pymod_name,(),())), *self._additional_imports.values()]
         imports  = ''.join(self._print(i) for i in imports)
 
         self.exit_scope()
 
         return '\n'.join(['#define PY_ARRAY_UNIQUE_SYMBOL CWRAPPER_ARRAY_API',
-                imports, decs, function_signatures, sep, class_defs, sep,
-                function_defs, exec_func, sep, method_def, sep, slots_def, sep,
+                f'#define {pymod_name.upper()}\n',
+                imports, decs, sep, class_defs, sep,
+                function_defs, sep, method_def, sep,
                 module_def, sep, init_func])
 
     def _print_PyClassDef(self, expr):
@@ -342,18 +385,10 @@ class CWrapperCodePrinter(CCodePrinter):
         name = self.scope.get_python_name(expr.name)
         docstring = self._print(LiteralString('\n'.join(expr.docstring.comments))) \
                     if expr.docstring else '""'
-        attributes = ''.join(self._print(Declare(a.dtype, a)) for a in expr.attributes)
-        class_code = (f"struct {struct_name} {{\n"
-                "    PyObject_HEAD\n"
-                + attributes +
-                "};\n")
 
         original_scope = expr.original_class.scope
         print_methods = expr.methods + (expr.new_func,) + expr.interfaces
         functions = '\n'.join(self._print(f) for f in print_methods)
-        sig_methods = expr.methods + (expr.new_func,) + tuple(f for i in expr.interfaces for f in i.functions) + \
-                      tuple(i.interface_func for i in expr.interfaces)
-        function_sigs = ''.join(self.function_signature(f)+';\n' for f in sig_methods)
         init_string = ''
         del_string = ''
         funcs = {}
@@ -400,7 +435,15 @@ class CWrapperCodePrinter(CCodePrinter):
                 f"    .tp_methods = {method_def_name},\n"
                 "};\n")
 
-        return class_code + '\n' + function_sigs + '\n' + method_def + '\n' + type_code + '\n' + functions
+        return method_def + '\n' + type_code + '\n' + functions
+
+    def _print_PyModInitFunc(self, expr):
+        decs = ''.join(self._print(d) for d in expr.declarations)
+        body = self._print(expr.body)
+        return ''.join([f'PyMODINIT_FUNC {expr.name}(void)\n{{\n',
+                decs,
+                body,
+                '}\n'])
 
     def _print_Allocate(self, expr):
         variable = expr.variable
@@ -424,6 +467,35 @@ class CWrapperCodePrinter(CCodePrinter):
             return f'{type_name}.tp_free({var_code});\n'
         else:
             return CCodePrinter._print_Deallocate(self, expr)
+
+    def _print_Declare(self, expr):
+        var = expr.variable
+        if isinstance(var.dtype, BindCPointer):
+            declaration_type = 'void*'
+
+            static = 'static ' if expr.static else ''
+            external = 'extern ' if expr.external else ''
+
+            variable = self._print(expr.variable.name)
+
+            if var.rank == 0:
+                return f'{static}{external}{declaration_type} {variable};\n'
+
+            size = var.shape[0]
+            if isinstance(size, LiteralInteger):
+                return f'{static}{external}{declaration_type} {variable}[{size}];\n'
+            else:
+                return f'{static}{external}{declaration_type}* {variable};\n'
+        else:
+            return CCodePrinter._print_Declare(self, expr)
+
+    def _print_IndexedElement(self, expr):
+        if isinstance(expr.base.class_type, CStackArray):
+            base = self._print(expr.base.name)
+            idxs = ''.join(f'[{self._print(a)}]' for a in expr.indices)
+            return f'{base}{idxs}'
+        else:
+            return CCodePrinter._print_IndexedElement(self, expr)
 
 def cwrappercode(expr, filename, target_language, assign_to=None, **settings):
     """Converts an expr to a string of c wrapper code